requirements.txt
# ============================================================================
# REQUIREMENTS.TXT - EVALUACIÓN 2: AGENTE INTELIGENTE EVERLAST
# ============================================================================
# Versión de Python requerida: >= 3.9
# Versiones 100% compatibles y probadas
# ============================================================================

# --- Core LangChain (BLOQUE COMPATIBLE) ---
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-openai==0.0.8
langchain-text-splitters==0.0.2

# --- OpenAI & Embeddings (VERSIONES ESTABLES) ---
openai==1.12.0
tiktoken==0.6.0

# --- Vector Store ---
faiss-cpu==1.8.0

# --- Agentes & Memoria ---
langchainhub==0.1.15

# --- Web Framework ---
streamlit==1.32.0

# --- Utilidades Core ---
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-core==2.16.3

# --- Data Processing ---
pandas==2.2.1
numpy==1.26.4

# --- Opcional: Evaluación ---
plotly==5.20.0
scikit-learn==1.4.2

# --- Opcional: LangSmith ---
langsmith==0.1.17

# --- Document Loaders ---
pypdf==4.1.0
python-docx==1.1.0
markdown==3.5.2

# --- Dependencias adicionales necesarias ---
requests==2.31.0
PyYAML==6.0.1
SQLAlchemy==2.0.28
aiohttp==3.9.3
dataclasses-json==0.6.4
jsonpatch==1.33
packaging==23.2
tenacity==8.2.3